feat：V1.8开发完成

+++ Yuan_club_subject/Yuan_club_domain/src/main/java/com/zhy/subject/domain/handler/subject/BriefTypeHandler.java
@@ -1,5 +1,7 @@
 package com.zhy.subject.domain.handler.subject;
 
+
+import com.zhy.subject.common.enums.IsDeleteFlagEnum;
 import com.zhy.subject.common.enums.SubjectInfoTypeEnum;
 import com.zhy.subject.domain.convert.BriefSubjectConverter;
 import com.zhy.subject.domain.entity.SubjectInfoBO;
@@ -7,20 +9,21 @@ import com.zhy.subject.domain.entity.SubjectOptionBO;
 import com.zhy.subject.infra.basic.entity.SubjectBrief;
 import com.zhy.subject.infra.basic.service.SubjectBriefService;
 import org.springframework.stereotype.Component;
-import com.zhy.subject.common.enums.IsDeleteFlagEnum;
 
 import javax.annotation.Resource;
 
 /**
- * @author Lenovo
- * @version 1.0
- * @description TODO
- * @date 17/1/2024 上午11:51
+ * 简答题目的策略类
+ * 
+ * @author: 随缘而愈
+ * @date: 2023/10/5
  */
 @Component
-public class BriefTypeHandler implements SubjectTypeHandler {
+public class BriefTypeHandler implements SubjectTypeHandler{
+
     @Resource
     private SubjectBriefService subjectBriefService;
+    
     @Override
     public SubjectInfoTypeEnum getHandlerType() {
         return SubjectInfoTypeEnum.BRIEF;
@@ -28,22 +31,19 @@ public class BriefTypeHandler implements SubjectTypeHandler {
 
     @Override
     public void add(SubjectInfoBO subjectInfoBO) {
-        //简答题目的插入
         SubjectBrief subjectBrief = BriefSubjectConverter.INSTANCE.convertBoToEntity(subjectInfoBO);
-        subjectBrief.setSubjectId(subjectInfoBO.getId());
+        subjectBrief.setSubjectId(subjectInfoBO.getId().intValue());
         subjectBrief.setIsDeleted(IsDeleteFlagEnum.UN_DELETED.getCode());
         subjectBriefService.insert(subjectBrief);
-
     }
 
     @Override
     public SubjectOptionBO query(int subjectId) {
         SubjectBrief subjectBrief = new SubjectBrief();
-        subjectBrief.setSubjectId(Long.valueOf(subjectId));
+        subjectBrief.setSubjectId(subjectId);
         SubjectBrief result = subjectBriefService.queryByCondition(subjectBrief);
         SubjectOptionBO subjectOptionBO = new SubjectOptionBO();
         subjectOptionBO.setSubjectAnswer(result.getSubjectAnswer());
         return subjectOptionBO;
-
     }
 }
